Add ideogram track

Add ideogram track

add_ideogram_track(cytoband = system.file("extdata", "cytoBand.txt",
    package = "circlize"), species = NULL, track = current_track() + 1)

Arguments

cytoband
Path of the cytoband file or a data frame that already contains cytoband data. Pass to read.cytoband.

species
Abbreviations of species. e.g. hg19 for human, mm10 for mouse. If this value is specified, the function will download cytoBand.txt.gz from UCSC ftp automatically. Pass to read.cytoband.

track
which track the ideogram is added in. By default it is the next track in the layout.

Details

A track which contains ideograms will be added to the plot.

The function tries to download cytoband file from UCSC ftp. If there is no cytoband file available for the species, there will be an error.

Value

No value is returned.

Examples

# There is no example
NULL

#> NULL